fix(ui): Migrate call sites to shared ItemCounts
Replace inline badge rendering in `ProductItemCounts`, `LastRunItemCounts`,
and `RepositoryRunsTable` with the shared `ItemCounts` component. Also fix
the missing search params for the links: each badge now links with
route-appropriate default sorting (descending severity for issues /
rule violations, descending rating for vulnerabilities).

Resolves #4571.
